Task 7 hibernate

+++ src/main/java/com/example/ConsoleApp.java
@@ -1,7 +1,7 @@
 package com.example;
 
 import com.example.localization.Localization;
-import com.example.model.Book;
+import com.example.entity.Book;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 import com.example.service.BookService;
@@ -40,34 +40,31 @@ public class ConsoleApp {
                         books.forEach(System.out::println);
                         break;
                     case "2":
-                        System.out.println(localization.getMessage("menu.book.id"));
-                        int id = scanner.nextInt();
-                        scanner.nextLine();
-//                        if (bookService.findBookById(id)) {
-//                            System.out.printf(localization.getMessage("menu.book.id.exist"));
-//                            break;
-//                        }
                         System.out.println(localization.getMessage("menu.book.enter.title"));
                         String title = scanner.nextLine();
                         System.out.println(localization.getMessage("menu.book.enter.author"));
                         String author = scanner.nextLine();
                         System.out.println(localization.getMessage("menu.book.enter.description"));
                         String description = scanner.nextLine();
-
-                        bookService.createBook(new Book(id, title, author, description));
+                        bookService.createBook(new Book(title, author, description));
                         System.out.println(localization.getMessage("menu.book.created"));
                         break;
                     case "3":
                         System.out.println(localization.getMessage("menu.book.id"));
                         int editId = scanner.nextInt();
                         scanner.nextLine();
+
+                        if (bookService.existsBookById(editId)) {
+                            System.out.println(localization.getMessage("menu.book.id.exist"));
+                            break;
+                        }
+
                         System.out.println(localization.getMessage("menu.book.enter.title"));
                         String newTitle = scanner.nextLine();
                         System.out.println(localization.getMessage("menu.book.enter.author"));
                         String newAuthor = scanner.nextLine();
                         System.out.println(localization.getMessage("menu.book.enter.description"));
                         String newDescription = scanner.nextLine();
-
                         bookService.updateBook(editId, new Book(editId, newTitle, newAuthor, newDescription));
                         System.out.println(localization.getMessage("menu.book.updated"));
                         break;
@@ -75,6 +72,11 @@ public class ConsoleApp {
                         System.out.println(localization.getMessage("menu.book.id"));
                         int deleteId = scanner.nextInt();
                         scanner.nextLine();
+
+                        if (bookService.existsBookById(deleteId)) {
+                            System.out.println(localization.getMessage("menu.book.id.exist"));
+                            break;
+                        }
 
                         bookService.deleteBook(deleteId);
                         System.out.println(localization.getMessage("menu.book.deleted"));
